♻️ refactor: separate displayButtons() to other file
